Updated Controller. Needs corrections and modifications.

+++ bookstore/src/main/java/app/bookstore/AuthorRepo.java
@@ -2,5 +2,6 @@ package app.bookstore;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface AuthorRepo extends CrudRepository<Author, Integer>
-{}
+public interface AuthorRepo extends CrudRepository<Author, Long> {
+
+}
