+++ bookstore/src/main/java/app/bookstore/UserRepo.java
@@ -2,5 +2,6 @@ package app.bookstore;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserRepo extends CrudRepository<User, Integer>
-{}
+public interface UserRepo extends CrudRepository<User, Long> {
+
+}
